# api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Fundly API
  version: 0.1.0
  description: |
    Kontrak REST JSON Fundly (PRD bagian 9). Semua endpoint berada di bawah `/api/v1`
    dan membutuhkan sesi (cookie HttpOnly `fundly_session`) kecuali ditandai publik.
    Permintaan yang mengubah data wajib membawa header `X-Requested-With: fundly`.
    Uang selalu bilangan bulat rupiah (tanpa desimal).
servers:
  - url: /api/v1
security:
  - sessionCookie: []
...
paths:
...
  /auth/google/callback:
    get:
      tags: [auth]
      summary: Callback OAuth Google (publik)
      operationId: googleCallback
      security: []
      parameters:
        - { name: code, in: query, required: true, schema: { type: string } }
        - { name: state, in: query, required: true, schema: { type: string } }
      responses:
        "302": { description: "Sesi dimulai, redirect ke aplikasi" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /me:
    get:
      tags: [me]
      summary: Profil dan pengaturan
      operationId: getMe
      responses:
        "200":
          description: Profil pengguna
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401": { $ref: "#/components/responses/Unauthorized" }
    patch:
      tags: [me]
      summary: Ubah profil, tema, atau password
      operationId: updateMe
      parameters:
        - $ref: "#/components/parameters/XRequestedWith"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMeRequest"
      responses:
        "200":
          description: Profil diperbarui
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/Unprocessable" }
    delete:
      tags: [me]
      summary: Hapus akun dan semua data secara permanen
      operationId: deleteMe
      parameters:
        - $ref: "#/components/parameters/XRequestedWith"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [confirm]
              properties:
                confirm:
                  type: string
                  const: HAPUS
                  description: Konfirmasi eksplisit
      responses:
        "204": { description: Akun dan semua data dihapus }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/Unprocessable" }
...
  /budgets:
    get:
      tags: [budgets]
      summary: Anggaran bulan beserta jumlah terpakai
      operationId: listBudgets
      parameters:
        - $ref: "#/components/parameters/MonthRequired"
      responses:
        "200":
          description: Daftar anggaran
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Budget" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /budgets/{category_id}:
    parameters:
      - name: category_id
        in: path
        required: true
        schema: { type: string, format: uuid }
    put:
      tags: [budgets]
      summary: Atur batas anggaran kategori untuk satu bulan
      operationId: putBudget
      parameters:
        - $ref: "#/components/parameters/MonthRequired"
        - $ref: "#/components/parameters/XRequestedWith"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [limit_amount]
              properties:
                limit_amount: { $ref: "#/components/schemas/Money" }
      responses:
        "200":
          description: Anggaran disimpan
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Budget" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/Unprocessable" }
    delete:
      tags: [budgets]
      summary: Hapus batas anggaran kategori untuk satu bulan
      operationId: deleteBudget
      parameters:
        - $ref: "#/components/parameters/MonthRequired"
        - $ref: "#/components/parameters/XRequestedWith"
      responses:
        "204": { description: Anggaran dihapus }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /export/transactions.csv:
    get:
      tags: [export]
      summary: Ekspor transaksi ke CSV (bulan tertentu, atau semua bila month kosong)
      operationId: exportTransactionsCsv
      parameters:
        - $ref: "#/components/parameters/Month"
      responses:
        "200":
          description: File CSV
          content:
            text/csv:
              schema: { type: string }
        "401": { $ref: "#/components/responses/Unauthorized" }
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: fundly_session
  parameters:
...
    Month:
      name: month
      in: query
      description: Bulan dalam format YYYY-MM (default bulan ini, zona Asia/Jakarta)
      schema: { type: string, pattern: "^[0-9]{4}-(0[1-9]|1[0-2])$" }
    MonthRequired:
      name: month
      in: query
      required: true
      description: Bulan dalam format YYYY-MM
      schema: { type: string, pattern: "^[0-9]{4}-(0[1-9]|1[0-2])$" }
    XRequestedWith:
      name: X-Requested-With
      in: header
      required: true
      description: Lapisan CSRF tambahan untuk permintaan yang mengubah data
      schema: { type: string, const: fundly }
  schemas:
    Money:
      type: integer
      format: int64
      minimum: 0
      description: Jumlah rupiah tanpa desimal
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
...
    User:
      type: object
      required: [id, email, display_name, theme, budget_auto_copy, has_password, has_google, created_at]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        display_name: { type: string }
        theme: { type: string, enum: [system, light, dark] }
        budget_auto_copy:
          type: boolean
          description: Anggaran bulan baru otomatis menyalin bulan lalu (F-06 KP3)
        has_password: { type: boolean }
        has_google: { type: boolean }
        created_at: { type: string, format: date-time }
...
    UpdateMeRequest:
      type: object
      properties:
        display_name: { type: string, maxLength: 60 }
        theme: { type: string, enum: [system, light, dark] }
        budget_auto_copy: { type: boolean }
        current_password: { type: string, maxLength: 128 }
        new_password: { type: string, minLength: 8, maxLength: 128 }
...
    Budget:
      type: object
      required: [category_id, month, limit_amount, spent, status]
      properties:
        category_id: { type: string, format: uuid }
        month: { type: string, pattern: "^[0-9]{4}-(0[1-9]|1[0-2])$" }
        limit_amount: { $ref: "#/components/schemas/Money" }
        spent: { $ref: "#/components/schemas/Money" }
        status:
          type: string
          enum: [safe, near, over]
          description: "safe < 80%, near ≥ 80%, over ≥ 100%"
  responses:
    BadRequest:
      description: Permintaan tidak valid
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Belum masuk atau sesi kedaluwarsa
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    NotFound:
      description: Tidak ditemukan
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    Unprocessable:
      description: Validasi gagal
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...