# api/internal/db/store/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package store

import (
	"time"

	"github.com/google/uuid"
)

type Budget struct {
	ID          uuid.UUID `json:"id"`
	UserID      uuid.UUID `json:"user_id"`
	CategoryID  uuid.UUID `json:"category_id"`
	Month       time.Time `json:"month"`
	LimitAmount int64     `json:"limit_amount"`
	CreatedAt   time.Time `json:"created_at"`
}

type Category struct {
	ID         uuid.UUID  `json:"id"`
	UserID     *uuid.UUID `json:"user_id"`
	Name       string     `json:"name"`
	Kind       string     `json:"kind"`
	Icon       string     `json:"icon"`
	ColorToken string     `json:"color_token"`
	SortOrder  int32      `json:"sort_order"`
	Hidden     bool       `json:"hidden"`
	CreatedAt  time.Time  `json:"created_at"`
}

type CategoryRule struct {
	ID         uuid.UUID  `json:"id"`
	UserID     *uuid.UUID `json:"user_id"`
	Keyword    string     `json:"keyword"`
	CategoryID uuid.UUID  `json:"category_id"`
	Priority   int32      `json:"priority"`
	CreatedAt  time.Time  `json:"created_at"`
}

type Session struct {
	ID         uuid.UUID `json:"id"`
	UserID     uuid.UUID `json:"user_id"`
	TokenHash  string    `json:"token_hash"`
	ExpiresAt  time.Time `json:"expires_at"`
	LastSeenAt time.Time `json:"last_seen_at"`
	CreatedAt  time.Time `json:"created_at"`
}

type Transaction struct {
	ID         uuid.UUID  `json:"id"`
	UserID     uuid.UUID  `json:"user_id"`
	WalletID   uuid.UUID  `json:"wallet_id"`
	CategoryID *uuid.UUID `json:"category_id"`
	Kind       string     `json:"kind"`
	Amount     int64      `json:"amount"`
	OccurredOn time.Time  `json:"occurred_on"`
	Merchant   *string    `json:"merchant"`
	Note       *string    `json:"note"`
	Source     string     `json:"source"`
	ClientID   *uuid.UUID `json:"client_id"`
	CreatedAt  time.Time  `json:"created_at"`
	UpdatedAt  time.Time  `json:"updated_at"`
	DeletedAt  *time.Time `json:"deleted_at"`
}

type User struct {
	ID             uuid.UUID `json:"id"`
	Email          string    `json:"email"`
	PasswordHash   *string   `json:"password_hash"`
	GoogleSub      *string   `json:"google_sub"`
	DisplayName    string    `json:"display_name"`
	Theme          string    `json:"theme"`
	CreatedAt      time.Time `json:"created_at"`
	BudgetAutoCopy bool      `json:"budget_auto_copy"`
}

type UserCategorySetting struct {
	UserID     uuid.UUID `json:"user_id"`
	CategoryID uuid.UUID `json:"category_id"`
	Hidden     bool      `json:"hidden"`
	SortOrder  *int32    `json:"sort_order"`
}

type Wallet struct {
	ID             uuid.UUID  `json:"id"`
	UserID         uuid.UUID  `json:"user_id"`
	Name           string     `json:"name"`
	Type           string     `json:"type"`
	Provider       *string    `json:"provider"`
	InitialBalance int64      `json:"initial_balance"`
	ArchivedAt     *time.Time `json:"archived_at"`
	CreatedAt      time.Time  `json:"created_at"`
}